git-svn-id: http://svn.seqan.de/seqan/trunk@1434 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/seqan/find_motif/find_motif_pms1.h
@@ -506,7 +506,7 @@ pms1(TStrings & result_set,
 	// STEP 2:
 	// count votes of l-mers (d-variants) and insert relevant d-variants into result set
 	// ----------------------------------------------------------------------------
-	int lower_limit = ceil(t*(model_type.threshold));
+	int lower_limit = (int) ceil(t*(model_type.threshold));
 
 	// count votes of relevant l-mers (d-variants)
 	std::vector<int>::iterator iter = result_vect.begin();
